migrate skills tests to bun test with real db
