+++ packages/server/__test__/skills/test-skill.md
