+++ packages/server/src/skills/built-in-skills.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, test } from 'vitest';
+import { describe, expect, test } from 'bun:test';
 
 import { getBuiltInSkillSource, loadBuiltInSkills } from '@/skills/built-in-skills.js';
 
@@ -6,7 +6,7 @@ describe('loadBuiltInSkills', () => {
   test('loads skills from markdown files', async () => {
     const skills = await loadBuiltInSkills([
       {
-        sourceUrl: new URL('../fixtures/skills/test-skill.md', import.meta.url),
+        sourceUrl: new URL('../../__test__/skills/test-skill.md', import.meta.url),
         bundledPath: 'unused-in-tests/test-skill.md',
       },
     ]);
@@ -22,7 +22,7 @@ describe('loadBuiltInSkills', () => {
 
   test('rejects duplicate skill names', async () => {
     const file = {
-      sourceUrl: new URL('../fixtures/skills/test-skill.md', import.meta.url),
+      sourceUrl: new URL('../../__test__/skills/test-skill.md', import.meta.url),
       bundledPath: 'unused-in-tests/test-skill.md',
     };
